+++ web/sites/default/files/civicrm/ext/authnet/Civi/Api4/AuthnetUtilities.php
@@ -1,0 +1,18 @@
+<?php
+namespace Civi\Api4;
+
+/**
+ * Authnet Utilities.
+ *
+ * Utilities provided by Authnet extension.
+ *
+ * @package Civi\Api4
+ */
+class AuthnetUtilities extends Generic\AbstractEntity {
+
+  public static function getFields() {
+    return new Generic\BasicGetFieldsAction(__CLASS__, __FUNCTION__, function() {
+      return [ ];
+    });
+  }
+}