+++ web/sites/default/files/civicrm/ext/authnet/authnetecheck.civix.php
@@ -89,7 +89,6 @@ class CRM_AuthNetEcheck_ExtensionUtil {
 
 use CRM_AuthNetEcheck_ExtensionUtil as E;
 
-pathload()->addSearchDir(__DIR__ . '/mixin/lib');
 spl_autoload_register('_authnetecheck_civix_class_loader', TRUE, TRUE);
 
 function _authnetecheck_civix_class_loader($class) {